refactor: sizeの指定による差分を分かりやすくした
