--- frontend/components/atoms/Spacer/HSpacer/index.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-
-import { HSpacer } from "../index";
-
-const meta: Meta<typeof HSpacer> = {
-  title: "atoms/HSpacer",
-  component: HSpacer,
-};
-
-export default meta;
-type Story = StoryObj<typeof HSpacer>;
-
-export const Default: Story = {
-  args: {},
-};
