--- frontend/components/atoms/Spacer/VSpacer/index.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-
-import { VSpacer } from "../index";
-
-const meta: Meta<typeof VSpacer> = {
-  title: "atoms/VSpacer",
-  component: VSpacer,
-};
-
-export default meta;
-type Story = StoryObj<typeof VSpacer>;
-
-export const Default: Story = {
-  args: {},
-};
